requirements.txt
# 简化版Airtest测试框架依赖
airtest
pocoui
